# lib/api-zod/src/generated/types/adminLoginResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Red-X Tube
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginResult {
  success: boolean;
}