# lib/api-zod/src/generated/types/siteStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Red-X Tube
 * OpenAPI spec version: 0.1.0
 */

export interface SiteStats {
  totalVideos: number;
  homeCount: number;
  freeCount: number;
  paidCount: number;
  totalViews: number;
  categoryCount: number;
  socialLinkCount: number;
}