# lib/api-zod/src/generated/types/socialLink.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Red-X Tube
 * OpenAPI spec version: 0.1.0
 */

export interface SocialLink {
  id: number;
  platform: string;
  url: string;
  displayName: string;
  iconName: string;
  /** @nullable */
  logoUrl?: string | null;
  sortOrder: number;
}